chore: Move experience <-> stars conversion to domain package
This is tightly coupled with our domain models, and not app logic per se
