+++ internal/app/milestone_test.go
@@ -2,7 +2,6 @@ package app
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"time"
 
@@ -27,50 +26,6 @@ func (m *mockMilestoneRepository) FindMilestoneAchievements(ctx context.Context,
 	return m.achievements, m.err
 }
 
-func TestStarsToExperience(t *testing.T) {
-	tests := []struct {
-		stars      int
-		experience int64
-	}{
-		// Test cases yoinked from prism
-		{1, 500},
-		{3, 3500},
-		{20, 87000},
-		{481, 2340000},
-		{1000, 4870000},
-		{1091, 5312000},
-	}
-
-	for _, tt := range tests {
-		t.Run(fmt.Sprintf("%d stars", tt.stars), func(t *testing.T) {
-			result := starsToExperience(tt.stars)
-			require.Equal(t, tt.experience, result)
-		})
-	}
-}
-
-func TestExperienceToStars(t *testing.T) {
-	tests := []struct {
-		experience int64
-		stars      int
-	}{
-		// Test cases yoinked from prism
-		{500, 1},
-		{3500, 3},
-		{87000, 20},
-		{2340000, 481},
-		{4870000, 1000},
-		{5312000, 1091},
-	}
-
-	for _, tt := range tests {
-		t.Run(fmt.Sprintf("%d experience", tt.experience), func(t *testing.T) {
-			result := experienceToStars(tt.experience)
-			require.Equal(t, tt.stars, result)
-		})
-	}
-}
-
 func TestFindMilestoneAchievements(t *testing.T) {
 	ctx := context.Background()
 	playerUUID := domaintest.NewUUID(t)
@@ -86,7 +41,7 @@ func TestFindMilestoneAchievements(t *testing.T) {
 		p3 := *domaintest.NewPlayerBuilder(playerUUID, time.Date(2024, time.March, 5, 15, 30, 0, 0, time.UTC)).WithExperience(487_550).Build()
 
 		starMilestones := []int64{1, 3, 100}
-		expMilestones := []int64{starsToExperience(1), starsToExperience(3), starsToExperience(100)}
+		expMilestones := []int64{domain.StarsToExperience(1), domain.StarsToExperience(3), domain.StarsToExperience(100)}
 
 		mockRepo := &mockMilestoneRepository{
 			t:                  t,
